specifying "main" breaks webpack setups
is it reasonable to omit the `main` entry altogether? 

when installing via npm and building with webpack, this breaks the webpack build as the `./assets/react/index.js` will be picked up as a module and bundled, causing broken dependencies (if you don't want react in the project).

BTW, thanx! unsemantic is awesome!